clean up spec and improve handling of git sources
