+++ pybio/spec/reader.py
@@ -1,6 +1,4 @@
-from pathlib import Path
-
-from pybio.spec import CommonSpec
+from pybio.spec.common import CommonSpec
 
 
 class ReaderSpec(CommonSpec):
